Make run-tests.sh behave differently locally: actually fix formatting and dont produce artifacts

+++ tests/run-tests.sh
@@ -1,21 +1,47 @@
 #! /bin/bash
 set -uo pipefail
 
+show_help() {
+  cat <<EOF
+usage: ./tests/run-tests.sh [--ci]
+
+  Note the script is ran from the repo root; Running without --ci
+  will run "ormolu --mode inplace" and fix offending files.
+
+  Options:
+    --ci      Instructs the script to save artifacts and
+              only checks syntax, instead of fixing it.
+EOF
+}
+
+ci=false
+while [[ $# -ne "0" ]]; do
+  case $1 in
+    --ci) ci=true; shift;;
+    *) show_help; exit 1;;
+  esac
+done
+
 ## Runs ormolu on all .hs files in a given project; sets the ormolu_ok
 ## variable to `false` in case ormolu fails. It also creates an artifact
 ## explaining the failure inside the tests folder.
 run_ormolu() {
   local proj=$1
   echo "Running ormolu on $proj"
-  ormolu --mode check $(find ./$proj -name '*.hs') 2> >(tee "tests/${proj}-ormolu.artifact")
-  local ormolu_res=$?
+  local ormolu_res=0
+  if $ci; then
+    ormolu --mode check $(find ./$proj -name '*.hs') 2> >(tee "tests/${proj}-ormolu.artifact")
+    ormolu_res=$?
+  else 
+    ormolu --mode inplace $(find ./$proj -name '*.hs') 
+    ormolu_res=$?
+  fi 
 
-  if [[ "$ormolu_res" -eq "0" ]]; then
+  if $ci && [[ "$ormolu_res" -eq "0" ]]; then
     rm "tests/${proj}-ormolu.artifact"
-    return 0
-  else
-    return 1
   fi
+
+  return $ormolu_res
 }
 
 ## Runs the cabal tests of a project; creates artifacts with potential failures.
@@ -25,16 +51,23 @@ run_cabal_test() {
 
   ## cd into the project, then generates the cabal file and run the necessary tests.
   pushd "$proj"
-  cabal run tests | tee "../tests/${proj}-cabal-test.artifact"
-  local cabal_res=$?
+
+  local cabal_res=0
+  if $ci; then
+    cabal run tests | tee "../tests/${proj}-cabal-test.artifact"
+    cabal_res=$?
+  else
+    cabal run tests
+    cabal_res=$?
+  fi
+  
   popd
 
-  if [[ "$cabal_res" -eq "0" ]]; then
+  if $ci && [[ "$cabal_res" -eq "0" ]]; then
     rm "tests/$proj-cabal-test.artifact"
-    return 0
-  else
-    return 1
   fi
+
+  return $cabal_res
 }
 
 ## Runs hlint on a project; creates artifacts with potential failures
@@ -42,15 +75,20 @@ run_hlint() {
   local proj=$1
   echo "Running 'hlint' on $proj"
 
-  hlint --hint="tests/hlint.yaml" ${proj} | tee "tests/${proj}-hlint.artifact"
-  local hlint_res=$?
-
-  if [[ "$hlint_res" -eq "0" ]]; then
-    rm "tests/$proj-hlint.artifact"
-    return 0
+  local hlint_res=0
+  if $ci; then
+    hlint --hint="tests/hlint.yaml" ${proj} | tee "tests/${proj}-hlint.artifact"
+    hlint_res=$?
   else
-    return 1
+    hlint --hint="tests/hlint.yaml" ${proj}
+    hlint_res=$?
   fi
+
+  if $ci && [[ "$hlint_res" -eq "0" ]]; then
+    rm "tests/$proj-hlint.artifact"
+  fi
+
+  return $hlint_res
 }
 
 projects=("cooked-validators" "examples")
